12 раздел 2 задание
